- refactor(admin): move accounts/drafts into _components/_lib; fix modal UX edge cases
