+++ src/app/admin/drafts/_lib/fightEditorUtils.ts
